fix: incorrect casting of top_p and frequency_penalty

+++ backend/open_webui/utils/payload.py
@@ -44,9 +44,9 @@ def apply_model_params_to_body(
 def apply_model_params_to_body_openai(params: dict, form_data: dict) -> dict:
     mappings = {
         "temperature": float,
-        "top_p": int,
+        "top_p": float,
         "max_tokens": int,
-        "frequency_penalty": int,
+        "frequency_penalty": float,
         "seed": lambda x: x,
         "stop": lambda x: [bytes(s, "utf-8").decode("unicode_escape") for s in x],
     }
